drv queries: move to dedicated file
